requirements.txt
numpy
opencv-python
Pillow
tensorflow
tensorflow-object-detection-api
protobuf